refactor: removed unused code and moved some constant to utils packages
removed unused code for cron jobs as now it's handled in a different package, moved some common constants to 'skyhelper-utils' package
