+++ src/handlers/index.ts
@@ -1,10 +1,8 @@
 export { default as logger } from "#handlers/logger";
 export { default as buildShardEmbed } from "#handlers/buildShardsEmbed";
-export { default as getShardStatus } from "#handlers/getShardStatus";
 export * from "#handlers/getDailyEventTimes";
 export { default as getTS } from "#handlers/getTS";
 export { default as dblStats } from "#handlers/dblStats";
 export { default as topggAutopost } from "#handlers/topggAutopost";
-export * from "#handlers/scheduledUpdates";
-export { default as scheduler } from "#handlers/scheduler";
+export * from "#handlers/sendDailyQuestReminder";
 export * from "#handlers/dailyQuestEmbed";
