+++ src/libs/constants/index.ts
@@ -1,6 +1,4 @@
-export { default as shardsInfo } from "#libs/constants/shardsInfo";
 export { default as eventTimes } from "#libs/constants/eventTimes";
-export { default as shardsTimelines } from "#libs/constants/shardsTimelines";
 export { default as spiritsData } from "#libs/constants/spiritsData";
 export * from "#libs/constants/seasonPrices";
 export { default as seasonsData } from "#libs/constants/seasonsData";
